conflics UIHomeScreen
